MAGE-1556 Implement SearchClientProvider
