+++ Service/DirectSendStrategy.php
@@ -3,12 +3,13 @@
 namespace Algolia\AlgoliaSearch\Service;
 
 use Algolia\AlgoliaSearch\Api\Data\IndexOptionsInterface;
+use Algolia\AlgoliaSearch\Api\SearchClientProviderInterface;
 use Algolia\AlgoliaSearch\Api\SendStrategyInterface;
 
 class DirectSendStrategy implements SendStrategyInterface
 {
     public function __construct(
-        private AlgoliaConnector $connector
+        private SearchClientProviderInterface $clientProvider
     ) {}
 
     public function isApplicable(int $storeId): bool
@@ -18,7 +19,7 @@ class DirectSendStrategy implements SendStrategyInterface
 
     public function send(IndexOptionsInterface $indexOptions, array $requests): array
     {
-        return $this->connector->getClient($indexOptions->getStoreId())
+        return $this->clientProvider->getClient($indexOptions->getStoreId())
             ->batch($indexOptions->getIndexName(), ['requests' => $requests]);
     }
 }
